Split schema into separate schema and metadata files
Schemae have both metadata (used for registering the artefact with
panopticon) and a schema (used to build the finder).

Previously, all this information was held in the one file -- the schema.
Now, it's held in two separate files -- `metadata/:slug.json` for the
data that should go on the artefact, and `schema/:slug.json` for the
schema iself.